I don't agree with this test
I find the behavior of the library where it sniffs the content of the string and do things differently very harmful.

If I'm adding a string, I'm adding a string. The string might look like a reserved JSON word but it's none of your business.

This behaviour is akin to treating o.element("foo","true") as o.element("foo",true);