Exposing JsonConfig used for property setting.
JsonConfig change the way the value is mangled, so this allows the
caller of the builder to tweak the behaviour.

(The reason I needed this change was because I use different default
values for some of the JsonConfig options, and this allows me to keep
tests happy with minimal modification. That said, I think this feature
makes sense on its own.)